fixed subst bug, and separated logic stuff
