+++ animated-transformer/src/lib/logic/relations.spec.ts
@@ -1,0 +1,28 @@
+/* Copyright 2023 Google LLC. All Rights Reserved.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+==============================================================================*/
+import { parseRel } from './relations';
+
+describe('generative_logic', () => {
+  beforeEach(() => {});
+
+  it('parseRel', () => {
+    const { relName, args } = parseRel('squishes _x _y:animal');
+    expect(relName).toEqual('squishes');
+    expect(args[0].varName).toEqual('_x');
+    expect(args[0].varType).toEqual('');
+    expect(args[1].varName).toEqual('_y');
+    expect(args[1].varType).toEqual('animal');
+  });
+});
